02_TOOLSETS/facility_registry/requirements.txt
click>=8.1
duckdb>=1.1
pandas>=2.1
geopandas>=0.14
rapidfuzz>=3.5
folium>=0.15
requests>=2.31
pyyaml>=6.0
tqdm>=4.66
pyarrow>=14.0
tabulate>=0.9
